@@ -2,19 +2,23 @@
   "minimum-stability": "RC",
   "repositories": [
     {
-      "type": "github",
-      "url": "https://github.com/ybelenko/slim-token-authentication"
+      "type": "vcs",
+      "url": "https://github.com/ybelenko/slim-token-authentication",
+      "no-api": true
     }
   ],
   "require": {
-    "php": ">=7.0 <=8.0",
-    "slim/slim": "3.*",
-    "dyorg/slim-token-authentication": "dev-per_route_apply"
+    "php": "^7.4 || ^8.0",
+    "slim/slim": "^4.5.0",
+    "dyorg/slim-token-authentication": "dev-slim4",
+    "ybelenko/openapi-data-mocker": "^1.0",
+    "ybelenko/openapi-data-mocker-server-middleware": "^1.0",
+    "slim/psr7": "^1.1.0",
+    "sentry/sdk": "^3.1"
   },
   "require-dev": {
-    "phpunit/phpunit": "^6.0 || ^7.0",
-    "overtrue/phplint": "^1.0",
-    "squizlabs/php_codesniffer": "^3.0"
+    "overtrue/phplint": "^2.0.2",
+    "squizlabs/php_codesniffer": "^3.5"
   },
   "autoload": {
     "psr-4": { "OpenAPIServer\\": [
@@ -22,16 +26,7 @@
       "src/"
     ]}
   },
-  "autoload-dev": {
-    "psr-4": { "OpenAPIServer\\": "test/" }
-  },
   "scripts": {
-    "test": [
-      "@test-apis",
-      "@test-models"
-    ],
-    "test-apis": "phpunit --testsuite Apis",
-    "test-models": "phpunit --testsuite Models",
     "phpcs": "phpcs",
     "phplint": "phplint ./ --exclude=vendor"
   }
