@@ -1,118 +1,76 @@
 <?php
 /**
  * CometVisu Manager backend
+ * PHP version 7.4
  *
- * PHP version 7
- *
- * @package OpenAPIServer\Api
+ * @package OpenAPIServer
  * @author  OpenAPI Generator team
- * @version 1.0.0
  * @link    https://github.com/openapitools/openapi-generator
  */
 
-require_once __DIR__ . '/vendor/autoload.php';
+/**
+ * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
+ * The version of the OpenAPI document: 1.0.0
+ * Generated by: https://github.com/openapitools/openapi-generator.git
+ */
 
+/**
+ * NOTE: This class is auto generated by the openapi generator program.
+ * https://github.com/openapitools/openapi-generator
+ */
+
+require_once __DIR__ . "/vendor/autoload.php";
+
+use OpenAPIServer\Container;
 use OpenAPIServer\SlimRouter;
+use OpenAPIServer\Middleware\SentryMiddleware;
 
+// load config file
 $config = [];
-
-/**
- * Slim Framework Settings
- * Ref: http://www.slimframework.com/docs/v3/objects/application.html#slim-default-settings
- */
-$config['settings'] = [
-    /**
-     * The protocol version used by the Response object. (Default: '1.1')
-     * Default: '1.1'
-     */
-    // 'httpVersion' => '1.1',
-
-    /**
-     * Size of each chunk read from the Response body when sending to the browser.
-     * Default: 4096
-     */
-    // 'responseChunkSize' => 4096,
-
-    /**
-     * If false, then no output buffering is enabled. If 'append' or 'prepend', then
-     * any echo or print statements are captured and are either appended or prepended
-     * to the Response returned from the route callable.
-     * Default: 'append'
-     */
-    // 'outputBuffering' => 'append',
-
-    /**
-     * When true, the route is calculated before any middleware is executed. This
-     * means that you can inspect route parameters in middleware if you need to.
-     * Default: false
-     */
-     'determineRouteBeforeAppMiddleware' => true,
-
-    /**
-     * When true, additional information about exceptions are displayed by the default
-     * error handler.
-     * Default: false
-     */
-     'displayErrorDetails' => true,
-
-    /**
-     * When true, Slim will add a Content-Length header to the response. If you are using
-     * a runtime analytics tool, such as New Relic, then this should be disabled.
-     * Default: true
-     */
-    // 'addContentLengthHeader' => true,
-
-    /**
-     * Filename for caching the FastRoute routes. Must be set to to a valid filename within
-     * a writeable directory. If the file does not exist, then it is created with the correct
-     * cache information on first run.
-     * Set to false to disable the FastRoute cache system.
-     * Default: false
-     */
-    // 'routerCacheFile' => false,
-];
-
-/**
- * Token Middleware 1.x Options
- * Options `header`, `regex`, `parameter`, `cookie`, `attribute`, `path`, `except`, `authenticator`
- * are handled by SlimRouter class. These options are ignored by app and they omitted from current
- * example.
- * Ref: https://github.com/dyorg/slim-token-authentication/tree/1.x
- */
-$config['tokenAuthenticationOptions'] = [
-    /**
-     * Tokens are essentially passwords. You should treat them as such and you should always
-     * use HTTPS. If the middleware detects insecure usage over HTTP it will return unathorized
-     * with a message Required HTTPS for token authentication. This rule is relaxed for requests
-     * on localhost. To allow insecure usage you must enable it manually by setting secure to
-     * false.
-     * Default: true
-     */
-    // 'secure' => true,
-
-    /**
-     * Alternatively you can list your development host to have relaxed security.
-     * Default: ['localhost', '127.0.0.1']
-     */
-    // 'relaxed' => ['localhost', '127.0.0.1'],
-
-    /**
-     * By default on ocurred a fail on authentication, is sent a response on json format with a
-     * message (`Invalid Token` or `Not found Token`) and with the token (if found), with status
-     * `401 Unauthorized`. You can customize it by setting a callable function on error option.
-     * Default: null
-     */
-    // 'error' => null,
-];
-
-$router = new SlimRouter($config);
+if (is_array($prodConfig = @include __DIR__ . "/config/prod/config.inc.php")) {
+    $config = $prodConfig;
+} elseif (
+    is_array($devConfig = @include __DIR__ . "/config/dev/config.inc.php")
+) {
+    $config = $devConfig;
+} else {
+    throw new InvalidArgumentException("Config file missed or broken.");
+}
+// Create Container using PHP-DI
+$container = new Container($config);
+$router = new SlimRouter($container);
 $app = $router->getSlimApp();
-$app->add(function($request, $response, $next) {
-  $response = $next($request, $response);
-  return $response
-    ->withHeader('Access-Control-Allow-Origin', '*')
-    ->withHeader('Access-Control-Allow-Headers', 'X-Requested-With, Content-Type, Accept, Origin, Authorization')
-    ->withHeader("Access-Control-Allow-Methods", 'GET, POST, PUT, DELETE, PATCH, OPTIONS');
-});
-//$router->initRoutes();
+$dir = $_SERVER["PHP_SELF"];
+$start = "/manager/index.php";
+$basePath = substr($dir, 0, mb_strpos($dir, $start) + mb_strlen($start));
+$app->setBasePath($basePath);
+
+// Parse json, form data and xml
+$app->addBodyParsingMiddleware();
+
+/**
+ * The routing middleware should be added before the ErrorMiddleware
+ * Otherwise exceptions thrown from it will not be handled
+ */
+$app->addRoutingMiddleware();
+
+$app->add(new SentryMiddleware);
+
+/**
+ * Add Error Handling Middleware
+ *
+ * @param bool $displayErrorDetails -> Should be set to false in production
+ * @param bool $logErrors -> Parameter is passed to the default ErrorHandler
+ * @param bool $logErrorDetails -> Display error details in error log
+ * which can be replaced by a callable of your choice.
+
+ * Note: This middleware should be added last. It will not handle any exceptions/errors
+ * for middleware added after it.
+ */
+$app->addErrorMiddleware(
+    $config["slimSettings"]["displayErrorDetails"] ?? false,
+    $config["slimSettings"]["logErrors"] ?? true,
+    $config["slimSettings"]["logErrorDetails"] ?? true
+);
+
 $app->run();
