@@ -1,17 +1,15 @@
 <?php
+
 /**
- * AbstractConfigApi
+ * CometVisu Manager backend
+ * PHP version 7.4
  *
- * PHP version 7
- *
- * @package OpenAPIServer\Api
+ * @package OpenAPIServer
  * @author  OpenAPI Generator team
  * @link    https://github.com/openapitools/openapi-generator
  */
 
 /**
- * CometVisu Manager backend
- *
  * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
  * The version of the OpenAPI document: 1.0.0
  * Generated by: https://github.com/openapitools/openapi-generator.git
@@ -27,7 +25,7 @@ namespace OpenAPIServer\Api;
 use Psr\Container\ContainerInterface;
 use Psr\Http\Message\ServerRequestInterface;
 use Psr\Http\Message\ResponseInterface;
-use Exception;
+use Slim\Exception\HttpNotImplementedException;
 
 /**
  * AbstractConfigApi Class Doc Comment
@@ -38,22 +36,20 @@ use Exception;
  */
 abstract class AbstractConfigApi
 {
-
     /**
-     * @var ContainerInterface Slim app container instance
+     * @var ContainerInterface|null Slim app container instance
      */
     protected $container;
 
     /**
      * Route Controller constructor receives container
      *
-     * @param ContainerInterface $container Slim app container instance
+     * @param ContainerInterface|null $container Slim app container instance
      */
-    public function __construct(ContainerInterface $container)
+    public function __construct(ContainerInterface $container = null)
     {
         $this->container = $container;
     }
-
 
     /**
      * POST createHiddenConfig
@@ -64,17 +60,19 @@ abstract class AbstractConfigApi
      * @param array|null             $args     Path arguments
      *
      * @return ResponseInterface
-     * @throws Exception to force implementation class to override this method
+     * @throws HttpNotImplementedException to force implementation class to override this method
      */
-    public function createHiddenConfig(ServerRequestInterface $request, ResponseInterface $response, array $args)
-    {
-        $section = $args['section'];
-        $key = $args['key'];
+    public function createHiddenConfig(
+        ServerRequestInterface $request,
+        ResponseInterface $response,
+        array $args
+    ) {
+        $section = $args["section"];
+        $key = $args["key"];
         $body = $request->getParsedBody();
-        $message = "How about implementing createHiddenConfig as a POST method in OpenAPIServer\Api\ConfigApi class?";
-        throw new Exception($message);
-
-        return $response->write($message)->withStatus(501);
+        $message =
+            "How about implementing createHiddenConfig as a POST method in OpenAPIServer\Api\ConfigApi class?";
+        throw new HttpNotImplementedException($request, $message);
     }
 
     /**
@@ -86,16 +84,18 @@ abstract class AbstractConfigApi
      * @param array|null             $args     Path arguments
      *
      * @return ResponseInterface
-     * @throws Exception to force implementation class to override this method
+     * @throws HttpNotImplementedException to force implementation class to override this method
      */
-    public function deleteHiddenConfig(ServerRequestInterface $request, ResponseInterface $response, array $args)
-    {
-        $section = $args['section'];
-        $key = $args['key'];
-        $message = "How about implementing deleteHiddenConfig as a DELETE method in OpenAPIServer\Api\ConfigApi class?";
-        throw new Exception($message);
-
-        return $response->write($message)->withStatus(501);
+    public function deleteHiddenConfig(
+        ServerRequestInterface $request,
+        ResponseInterface $response,
+        array $args
+    ) {
+        $section = $args["section"];
+        $key = $args["key"];
+        $message =
+            "How about implementing deleteHiddenConfig as a DELETE method in OpenAPIServer\Api\ConfigApi class?";
+        throw new HttpNotImplementedException($request, $message);
     }
 
     /**
@@ -108,16 +108,18 @@ abstract class AbstractConfigApi
      * @param array|null             $args     Path arguments
      *
      * @return ResponseInterface
-     * @throws Exception to force implementation class to override this method
+     * @throws HttpNotImplementedException to force implementation class to override this method
      */
-    public function getHiddenConfig(ServerRequestInterface $request, ResponseInterface $response, array $args)
-    {
-        $section = $args['section'];
-        $key = $args['key'];
-        $message = "How about implementing getHiddenConfig as a GET method in OpenAPIServer\Api\ConfigApi class?";
-        throw new Exception($message);
-
-        return $response->write($message)->withStatus(501);
+    public function getHiddenConfig(
+        ServerRequestInterface $request,
+        ResponseInterface $response,
+        array $args
+    ) {
+        $section = $args["section"];
+        $key = $args["key"];
+        $message =
+            "How about implementing getHiddenConfig as a GET method in OpenAPIServer\Api\ConfigApi class?";
+        throw new HttpNotImplementedException($request, $message);
     }
 
     /**
@@ -129,15 +131,17 @@ abstract class AbstractConfigApi
      * @param array|null             $args     Path arguments
      *
      * @return ResponseInterface
-     * @throws Exception to force implementation class to override this method
+     * @throws HttpNotImplementedException to force implementation class to override this method
      */
-    public function saveHiddenConfig(ServerRequestInterface $request, ResponseInterface $response, array $args)
-    {
+    public function saveHiddenConfig(
+        ServerRequestInterface $request,
+        ResponseInterface $response,
+        array $args
+    ) {
         $body = $request->getParsedBody();
-        $message = "How about implementing saveHiddenConfig as a PUT method in OpenAPIServer\Api\ConfigApi class?";
-        throw new Exception($message);
-
-        return $response->write($message)->withStatus(501);
+        $message =
+            "How about implementing saveHiddenConfig as a PUT method in OpenAPIServer\Api\ConfigApi class?";
+        throw new HttpNotImplementedException($request, $message);
     }
 
     /**
@@ -149,16 +153,18 @@ abstract class AbstractConfigApi
      * @param array|null             $args     Path arguments
      *
      * @return ResponseInterface
-     * @throws Exception to force implementation class to override this method
+     * @throws HttpNotImplementedException to force implementation class to override this method
      */
-    public function updateHiddenConfig(ServerRequestInterface $request, ResponseInterface $response, array $args)
-    {
-        $section = $args['section'];
-        $key = $args['key'];
+    public function updateHiddenConfig(
+        ServerRequestInterface $request,
+        ResponseInterface $response,
+        array $args
+    ) {
+        $section = $args["section"];
+        $key = $args["key"];
         $body = $request->getParsedBody();
-        $message = "How about implementing updateHiddenConfig as a PUT method in OpenAPIServer\Api\ConfigApi class?";
-        throw new Exception($message);
-
-        return $response->write($message)->withStatus(501);
+        $message =
+            "How about implementing updateHiddenConfig as a PUT method in OpenAPIServer\Api\ConfigApi class?";
+        throw new HttpNotImplementedException($request, $message);
     }
 }
