@@ -1,17 +1,15 @@
 <?php
+
 /**
- * AbstractDataproviderApi
+ * CometVisu Manager backend
+ * PHP version 7.4
  *
- * PHP version 7
- *
- * @package OpenAPIServer\Api
+ * @package OpenAPIServer
  * @author  OpenAPI Generator team
  * @link    https://github.com/openapitools/openapi-generator
  */
 
 /**
- * CometVisu Manager backend
- *
  * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
  * The version of the OpenAPI document: 1.0.0
  * Generated by: https://github.com/openapitools/openapi-generator.git
@@ -27,7 +25,7 @@ namespace OpenAPIServer\Api;
 use Psr\Container\ContainerInterface;
 use Psr\Http\Message\ServerRequestInterface;
 use Psr\Http\Message\ResponseInterface;
-use Exception;
+use Slim\Exception\HttpNotImplementedException;
 
 /**
  * AbstractDataproviderApi Class Doc Comment
@@ -38,22 +36,20 @@ use Exception;
  */
 abstract class AbstractDataproviderApi
 {
-
     /**
-     * @var ContainerInterface Slim app container instance
+     * @var ContainerInterface|null Slim app container instance
      */
     protected $container;
 
     /**
      * Route Controller constructor receives container
      *
-     * @param ContainerInterface $container Slim app container instance
+     * @param ContainerInterface|null $container Slim app container instance
      */
-    public function __construct(ContainerInterface $container)
+    public function __construct(ContainerInterface $container = null)
     {
         $this->container = $container;
     }
-
 
     /**
      * GET getAddresses
@@ -65,14 +61,16 @@ abstract class AbstractDataproviderApi
      * @param array|null             $args     Path arguments
      *
      * @return ResponseInterface
-     * @throws Exception to force implementation class to override this method
+     * @throws HttpNotImplementedException to force implementation class to override this method
      */
-    public function getAddresses(ServerRequestInterface $request, ResponseInterface $response, array $args)
-    {
-        $message = "How about implementing getAddresses as a GET method in OpenAPIServer\Api\DataproviderApi class?";
-        throw new Exception($message);
-
-        return $response->write($message)->withStatus(501);
+    public function getAddresses(
+        ServerRequestInterface $request,
+        ResponseInterface $response,
+        array $args
+    ) {
+        $message =
+            "How about implementing getAddresses as a GET method in OpenAPIServer\Api\DataproviderApi class?";
+        throw new HttpNotImplementedException($request, $message);
     }
 
     /**
@@ -85,14 +83,16 @@ abstract class AbstractDataproviderApi
      * @param array|null             $args     Path arguments
      *
      * @return ResponseInterface
-     * @throws Exception to force implementation class to override this method
+     * @throws HttpNotImplementedException to force implementation class to override this method
      */
-    public function getDesigns(ServerRequestInterface $request, ResponseInterface $response, array $args)
-    {
-        $message = "How about implementing getDesigns as a GET method in OpenAPIServer\Api\DataproviderApi class?";
-        throw new Exception($message);
-
-        return $response->write($message)->withStatus(501);
+    public function getDesigns(
+        ServerRequestInterface $request,
+        ResponseInterface $response,
+        array $args
+    ) {
+        $message =
+            "How about implementing getDesigns as a GET method in OpenAPIServer\Api\DataproviderApi class?";
+        throw new HttpNotImplementedException($request, $message);
     }
 
     /**
@@ -105,17 +105,21 @@ abstract class AbstractDataproviderApi
      * @param array|null             $args     Path arguments
      *
      * @return ResponseInterface
-     * @throws Exception to force implementation class to override this method
+     * @throws HttpNotImplementedException to force implementation class to override this method
      */
-    public function getInfluxDBFields(ServerRequestInterface $request, ResponseInterface $response, array $args)
-    {
+    public function getInfluxDBFields(
+        ServerRequestInterface $request,
+        ResponseInterface $response,
+        array $args
+    ) {
         $queryParams = $request->getQueryParams();
-        $auth = $request->getQueryParam('auth');
-        $measurement = $request->getQueryParam('measurement');
-        $message = "How about implementing getInfluxDBFields as a GET method in OpenAPIServer\Api\DataproviderApi class?";
-        throw new Exception($message);
-
-        return $response->write($message)->withStatus(501);
+        $auth = key_exists("auth", $queryParams) ? $queryParams["auth"] : null;
+        $measurement = key_exists("measurement", $queryParams)
+            ? $queryParams["measurement"]
+            : null;
+        $message =
+            "How about implementing getInfluxDBFields as a GET method in OpenAPIServer\Api\DataproviderApi class?";
+        throw new HttpNotImplementedException($request, $message);
     }
 
     /**
@@ -128,17 +132,21 @@ abstract class AbstractDataproviderApi
      * @param array|null             $args     Path arguments
      *
      * @return ResponseInterface
-     * @throws Exception to force implementation class to override this method
+     * @throws HttpNotImplementedException to force implementation class to override this method
      */
-    public function getInfluxDBTags(ServerRequestInterface $request, ResponseInterface $response, array $args)
-    {
+    public function getInfluxDBTags(
+        ServerRequestInterface $request,
+        ResponseInterface $response,
+        array $args
+    ) {
         $queryParams = $request->getQueryParams();
-        $auth = $request->getQueryParam('auth');
-        $measurement = $request->getQueryParam('measurement');
-        $message = "How about implementing getInfluxDBTags as a GET method in OpenAPIServer\Api\DataproviderApi class?";
-        throw new Exception($message);
-
-        return $response->write($message)->withStatus(501);
+        $auth = key_exists("auth", $queryParams) ? $queryParams["auth"] : null;
+        $measurement = key_exists("measurement", $queryParams)
+            ? $queryParams["measurement"]
+            : null;
+        $message =
+            "How about implementing getInfluxDBTags as a GET method in OpenAPIServer\Api\DataproviderApi class?";
+        throw new HttpNotImplementedException($request, $message);
     }
 
     /**
@@ -151,16 +159,18 @@ abstract class AbstractDataproviderApi
      * @param array|null             $args     Path arguments
      *
      * @return ResponseInterface
-     * @throws Exception to force implementation class to override this method
+     * @throws HttpNotImplementedException to force implementation class to override this method
      */
-    public function getInfluxDBs(ServerRequestInterface $request, ResponseInterface $response, array $args)
-    {
+    public function getInfluxDBs(
+        ServerRequestInterface $request,
+        ResponseInterface $response,
+        array $args
+    ) {
         $queryParams = $request->getQueryParams();
-        $auth = $request->getQueryParam('auth');
-        $message = "How about implementing getInfluxDBs as a GET method in OpenAPIServer\Api\DataproviderApi class?";
-        throw new Exception($message);
-
-        return $response->write($message)->withStatus(501);
+        $auth = key_exists("auth", $queryParams) ? $queryParams["auth"] : null;
+        $message =
+            "How about implementing getInfluxDBs as a GET method in OpenAPIServer\Api\DataproviderApi class?";
+        throw new HttpNotImplementedException($request, $message);
     }
 
     /**
@@ -173,13 +183,15 @@ abstract class AbstractDataproviderApi
      * @param array|null             $args     Path arguments
      *
      * @return ResponseInterface
-     * @throws Exception to force implementation class to override this method
+     * @throws HttpNotImplementedException to force implementation class to override this method
      */
-    public function getRRDs(ServerRequestInterface $request, ResponseInterface $response, array $args)
-    {
-        $message = "How about implementing getRRDs as a GET method in OpenAPIServer\Api\DataproviderApi class?";
-        throw new Exception($message);
-
-        return $response->write($message)->withStatus(501);
+    public function getRRDs(
+        ServerRequestInterface $request,
+        ResponseInterface $response,
+        array $args
+    ) {
+        $message =
+            "How about implementing getRRDs as a GET method in OpenAPIServer\Api\DataproviderApi class?";
+        throw new HttpNotImplementedException($request, $message);
     }
 }
