@@ -1,17 +1,15 @@
 <?php
+
 /**
- * AbstractFsApi
+ * CometVisu Manager backend
+ * PHP version 7.4
  *
- * PHP version 7
- *
- * @package OpenAPIServer\Api
+ * @package OpenAPIServer
  * @author  OpenAPI Generator team
  * @link    https://github.com/openapitools/openapi-generator
  */
 
 /**
- * CometVisu Manager backend
- *
  * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
  * The version of the OpenAPI document: 1.0.0
  * Generated by: https://github.com/openapitools/openapi-generator.git
@@ -27,7 +25,7 @@ namespace OpenAPIServer\Api;
 use Psr\Container\ContainerInterface;
 use Psr\Http\Message\ServerRequestInterface;
 use Psr\Http\Message\ResponseInterface;
-use Exception;
+use Slim\Exception\HttpNotImplementedException;
 
 /**
  * AbstractFsApi Class Doc Comment
@@ -38,22 +36,20 @@ use Exception;
  */
 abstract class AbstractFsApi
 {
-
     /**
-     * @var ContainerInterface Slim app container instance
+     * @var ContainerInterface|null Slim app container instance
      */
     protected $container;
 
     /**
      * Route Controller constructor receives container
      *
-     * @param ContainerInterface $container Slim app container instance
+     * @param ContainerInterface|null $container Slim app container instance
      */
-    public function __construct(ContainerInterface $container)
+    public function __construct(ContainerInterface $container = null)
     {
         $this->container = $container;
     }
-
 
     /**
      * GET checkEnvironment
@@ -65,14 +61,16 @@ abstract class AbstractFsApi
      * @param array|null             $args     Path arguments
      *
      * @return ResponseInterface
-     * @throws Exception to force implementation class to override this method
+     * @throws HttpNotImplementedException to force implementation class to override this method
      */
-    public function checkEnvironment(ServerRequestInterface $request, ResponseInterface $response, array $args)
-    {
-        $message = "How about implementing checkEnvironment as a GET method in OpenAPIServer\Api\FsApi class?";
-        throw new Exception($message);
-
-        return $response->write($message)->withStatus(501);
+    public function checkEnvironment(
+        ServerRequestInterface $request,
+        ResponseInterface $response,
+        array $args
+    ) {
+        $message =
+            "How about implementing checkEnvironment as a GET method in OpenAPIServer\Api\FsApi class?";
+        throw new HttpNotImplementedException($request, $message);
     }
 
     /**
@@ -84,21 +82,26 @@ abstract class AbstractFsApi
      * @param array|null             $args     Path arguments
      *
      * @return ResponseInterface
-     * @throws Exception to force implementation class to override this method
+     * @throws HttpNotImplementedException to force implementation class to override this method
      */
-    public function create(ServerRequestInterface $request, ResponseInterface $response, array $args)
-    {
+    public function create(
+        ServerRequestInterface $request,
+        ResponseInterface $response,
+        array $args
+    ) {
         $queryParams = $request->getQueryParams();
-        $path = $request->getQueryParam('path');
-        $hash = $request->getQueryParam('hash');
-        $type = $request->getQueryParam('type');
-        $file = (key_exists('file', $request->getUploadedFiles())) ? $request->getUploadedFiles()['file'] : null;
-        $filename = $request->getParsedBodyParam('filename');
-        $force = $request->getParsedBodyParam('force');
-        $message = "How about implementing create as a POST method in OpenAPIServer\Api\FsApi class?";
-        throw new Exception($message);
-
-        return $response->write($message)->withStatus(501);
+        $path = key_exists("path", $queryParams) ? $queryParams["path"] : null;
+        $hash = key_exists("hash", $queryParams) ? $queryParams["hash"] : null;
+        $type = key_exists("type", $queryParams) ? $queryParams["type"] : null;
+        $body = $request->getParsedBody();
+        $file = key_exists("file", $request->getUploadedFiles())
+            ? $request->getUploadedFiles()["file"]
+            : null;
+        $filename = isset($body["filename"]) ? $body["filename"] : null;
+        $force = isset($body["force"]) ? $body["force"] : null;
+        $message =
+            "How about implementing create as a POST method in OpenAPIServer\Api\FsApi class?";
+        throw new HttpNotImplementedException($request, $message);
     }
 
     /**
@@ -110,18 +113,22 @@ abstract class AbstractFsApi
      * @param array|null             $args     Path arguments
      *
      * @return ResponseInterface
-     * @throws Exception to force implementation class to override this method
+     * @throws HttpNotImplementedException to force implementation class to override this method
      */
-    public function delete(ServerRequestInterface $request, ResponseInterface $response, array $args)
-    {
+    public function delete(
+        ServerRequestInterface $request,
+        ResponseInterface $response,
+        array $args
+    ) {
         $queryParams = $request->getQueryParams();
-        $path = $request->getQueryParam('path');
-        $hash = $request->getQueryParam('hash');
-        $force = $request->getQueryParam('force');
-        $message = "How about implementing delete as a DELETE method in OpenAPIServer\Api\FsApi class?";
-        throw new Exception($message);
-
-        return $response->write($message)->withStatus(501);
+        $path = key_exists("path", $queryParams) ? $queryParams["path"] : null;
+        $hash = key_exists("hash", $queryParams) ? $queryParams["hash"] : null;
+        $force = key_exists("force", $queryParams)
+            ? $queryParams["force"]
+            : null;
+        $message =
+            "How about implementing delete as a DELETE method in OpenAPIServer\Api\FsApi class?";
+        throw new HttpNotImplementedException($request, $message);
     }
 
     /**
@@ -133,17 +140,21 @@ abstract class AbstractFsApi
      * @param array|null             $args     Path arguments
      *
      * @return ResponseInterface
-     * @throws Exception to force implementation class to override this method
+     * @throws HttpNotImplementedException to force implementation class to override this method
      */
-    public function move(ServerRequestInterface $request, ResponseInterface $response, array $args)
-    {
+    public function move(
+        ServerRequestInterface $request,
+        ResponseInterface $response,
+        array $args
+    ) {
         $queryParams = $request->getQueryParams();
-        $src = $request->getQueryParam('src');
-        $target = $request->getQueryParam('target');
-        $message = "How about implementing move as a PUT method in OpenAPIServer\Api\FsApi class?";
-        throw new Exception($message);
-
-        return $response->write($message)->withStatus(501);
+        $src = key_exists("src", $queryParams) ? $queryParams["src"] : null;
+        $target = key_exists("target", $queryParams)
+            ? $queryParams["target"]
+            : null;
+        $message =
+            "How about implementing move as a PUT method in OpenAPIServer\Api\FsApi class?";
+        throw new HttpNotImplementedException($request, $message);
     }
 
     /**
@@ -157,19 +168,25 @@ abstract class AbstractFsApi
      * @param array|null             $args     Path arguments
      *
      * @return ResponseInterface
-     * @throws Exception to force implementation class to override this method
+     * @throws HttpNotImplementedException to force implementation class to override this method
      */
-    public function read(ServerRequestInterface $request, ResponseInterface $response, array $args)
-    {
+    public function read(
+        ServerRequestInterface $request,
+        ResponseInterface $response,
+        array $args
+    ) {
         $queryParams = $request->getQueryParams();
-        $path = $request->getQueryParam('path');
-        $hash = $request->getQueryParam('hash');
-        $download = $request->getQueryParam('download');
-        $recursive = $request->getQueryParam('recursive');
-        $message = "How about implementing read as a GET method in OpenAPIServer\Api\FsApi class?";
-        throw new Exception($message);
-
-        return $response->write($message)->withStatus(501);
+        $path = key_exists("path", $queryParams) ? $queryParams["path"] : null;
+        $hash = key_exists("hash", $queryParams) ? $queryParams["hash"] : null;
+        $download = key_exists("download", $queryParams)
+            ? $queryParams["download"]
+            : null;
+        $recursive = key_exists("recursive", $queryParams)
+            ? $queryParams["recursive"]
+            : null;
+        $message =
+            "How about implementing read as a GET method in OpenAPIServer\Api\FsApi class?";
+        throw new HttpNotImplementedException($request, $message);
     }
 
     /**
@@ -181,17 +198,19 @@ abstract class AbstractFsApi
      * @param array|null             $args     Path arguments
      *
      * @return ResponseInterface
-     * @throws Exception to force implementation class to override this method
+     * @throws HttpNotImplementedException to force implementation class to override this method
      */
-    public function update(ServerRequestInterface $request, ResponseInterface $response, array $args)
-    {
+    public function update(
+        ServerRequestInterface $request,
+        ResponseInterface $response,
+        array $args
+    ) {
         $queryParams = $request->getQueryParams();
-        $path = $request->getQueryParam('path');
-        $hash = $request->getQueryParam('hash');
+        $path = key_exists("path", $queryParams) ? $queryParams["path"] : null;
+        $hash = key_exists("hash", $queryParams) ? $queryParams["hash"] : null;
         $body = $request->getParsedBody();
-        $message = "How about implementing update as a PUT method in OpenAPIServer\Api\FsApi class?";
-        throw new Exception($message);
-
-        return $response->write($message)->withStatus(501);
+        $message =
+            "How about implementing update as a PUT method in OpenAPIServer\Api\FsApi class?";
+        throw new HttpNotImplementedException($request, $message);
     }
 }
